chore: :tada: Setup project structure and initial components
I removed the extra files in the src folder and created the following directories: assets, components, context, and constants.
I set up the initial structure for the Contact List app by adding the `App.jsx` file with imports for the Header and Contacts components.
Additionally, I created the `global.css` for global styles, added an `inputs.js` file in the constants folder, and implemented the Header and Contacts components with their respective styles in CSS modules.