feat: :sparkles: I added ContactsList component with styling to display contacts
I created the `ContactsList.jsx` component to display the contact list and styled it using `ContactsList.module.css`.
The component uses `useContext` to access contacts from `ContactContext`.
If there are contacts, they are rendered through the `ContactItem` component. Otherwise, a message is shown to indicate no contacts were found.
The styles provide layout and presentation for the contact list and the message.